More organizational cleanup
